make some more tests, add a service layer and modified controller
