+++ Project2/project2/src/main/java/com/revature/Project2Application.java
@@ -1,6 +1,5 @@
 package com.revature;
 
-import com.revature.controllers.AuthController;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.domain.EntityScan;
