+++ Project2/project2/src/main/java/com/revature/tests/AccountsTypeController.java
@@ -1,6 +1,6 @@
-package com.revature.controllers;
+package com.revature.tests;
 
-import com.revature.daos.InnerTransferLogDAO;
+import com.revature.daos.AccountTypeDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -9,11 +9,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @CrossOrigin
 @RequestMapping()//TODO
-public class InnerTransferLogController {
-    private InnerTransferLogDAO itlDAO;
-
+public class AccountsTypeController {
+    private AccountTypeDAO atDAO;
     @Autowired
-    public InnerTransferLogController(InnerTransferLogDAO itlDAO) {
-        this.itlDAO = itlDAO;
+    public AccountsTypeController(AccountTypeDAO atDAO) {
+        this.atDAO = atDAO;
     }
 }
