+++ Project2/project2/src/main/java/com/revature/tests/RequestsController.java
@@ -1,6 +1,6 @@
-package com.revature.controllers;
+package com.revature.tests;
 
-import com.revature.daos.StatusDAO;
+import com.revature.daos.RequestsDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -9,10 +9,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @CrossOrigin
 @RequestMapping()//TODO
-public class StatusController {
-    private StatusDAO sDAO;
+public class RequestsController {
+    private RequestsDAO rDAO;
+
     @Autowired
-    public StatusController(StatusDAO sDAO) {
-        this.sDAO = sDAO;
+    public RequestsController(RequestsDAO rDAO) {
+        this.rDAO = rDAO;
     }
 }
