+++ Project2/project2/src/main/java/com/revature/tests/StatusController.java
@@ -1,6 +1,6 @@
-package com.revature.controllers;
+package com.revature.tests;
 
-import com.revature.daos.RequestsDAO;
+import com.revature.daos.StatusDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -9,11 +9,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @CrossOrigin
 @RequestMapping()//TODO
-public class RequestsController {
-    private RequestsDAO rDAO;
-
+public class StatusController {
+    private StatusDAO sDAO;
     @Autowired
-    public RequestsController(RequestsDAO rDAO) {
-        this.rDAO = rDAO;
+    public StatusController(StatusDAO sDAO) {
+        this.sDAO = sDAO;
     }
 }
