+++ Project2/project2/src/main/java/com/revature/tests/TransactionTypesController.java
@@ -1,13 +1,10 @@
-package com.revature.controllers;
+package com.revature.tests;
 
 import com.revature.daos.TransactionTypesDAO;
-import com.revature.models.TransactionType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.util.Optional;
 
 //we do not need this class
 @RestController
