+++ Project2/project2/src/test/java/com/revature/tests/TransactionsTest.java
@@ -1,4 +1,4 @@
-package com.revature.controllers;
+package com.revature.tests;
 
 
 public class TransactionsTest {
